+++ Tests/MapboxMapsTests/Integration Tests/FeatureQueryingTest.swift
@@ -212,7 +212,7 @@ internal class FeatureQueryingTest: MapViewIntegrationTestCase {
         geoJSONClusterSource.data = .featureCollection(FeatureCollection(features: features))
         geoJSONClusterSource.cluster = true
 
-        var geoJSONLayer = CircleLayer(id: "cluster-layer", source: clusterSourceID)
+        let geoJSONLayer = CircleLayer(id: "cluster-layer", source: clusterSourceID)
 
         didFinishLoadingStyle = { mapView in